pyproject.toml
[tool.poetry]
name = "maxkb"
version = "0.1.0"
description = "Intelligent Knowledge Base"
authors = ["shaohuzhang1 <shaohu.zhang@fit2cloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.1.13"
djangorestframework = "3.14.0"
drf-yasg = "1.21.7"
django-filter = "23.2"
langchain = "^0.1.11"
psycopg2-binary = "2.9.9"
jieba = "^0.42.1"
diskcache = "^5.6.3"
pillow = "^10.2.0"
filetype = "^1.2.0"
torch = "^2.2.1"
sentence-transformers = "^2.2.2"
blinker = "^1.6.3"
openai = "^1.13.3"
tiktoken = "^0.5.1"
qianfan = "^0.3.6.1"
pycryptodome = "^3.19.0"
beautifulsoup4 = "^4.12.2"
html2text = "^2024.2.26"
langchain-openai = "^0.0.8"
django-ipware = "^6.0.4"
django-apscheduler = "^0.6.2"
pymupdf = "1.24.1"
python-docx = "^1.1.0"
xlwt = "^1.3.0"
dashscope = "^1.17.0"
zhipuai = "^2.0.1"
httpx = "^0.27.0"
httpx-sse = "^0.4.0"
websocket-client = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"